remove simple build feature
STONEBLD-2612
