+++ controllers/component_build_controller_common.go
@@ -20,9 +20,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/url"
-	"path/filepath"
-	"sort"
 	"strings"
 
 	appstudiov1alpha1 "github.com/konflux-ci/application-api/api/v1alpha1"
@@ -33,7 +30,6 @@ import (
 	"gopkg.in/yaml.v2"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
@@ -47,54 +43,6 @@ type BuildPipeline struct {
 type pipelineConfig struct {
 	DefaultPipelineName string          `yaml:"default-pipeline-name"`
 	Pipelines           []BuildPipeline `yaml:"pipelines"`
-}
-
-// getGitProvider returns git provider name based on the repository url, e.g. github, gitlab, etc or git-privider annotation
-func getGitProvider(component appstudiov1alpha1.Component) (string, error) {
-	allowedGitProviders := map[string]bool{"github": true, "gitlab": true, "bitbucket": true}
-	gitProvider := ""
-
-	if component.Spec.Source.GitSource == nil {
-		err := fmt.Errorf("git source URL is not set for %s Component in %s namespace", component.Name, component.Namespace)
-		return "", err
-	}
-	sourceUrl := component.Spec.Source.GitSource.URL
-
-	if strings.HasPrefix(sourceUrl, "git@") {
-		// git@github.com:redhat-appstudio/application-service.git
-		sourceUrl = strings.TrimPrefix(sourceUrl, "git@")
-		host := strings.Split(sourceUrl, ":")[0]
-		gitProvider = strings.Split(host, ".")[0]
-	} else {
-		// https://github.com/redhat-appstudio/application-service
-		u, err := url.Parse(sourceUrl)
-		if err != nil {
-			return "", err
-		}
-		uParts := strings.Split(u.Hostname(), ".")
-		if len(uParts) == 1 {
-			gitProvider = uParts[0]
-		} else {
-			gitProvider = uParts[len(uParts)-2]
-		}
-	}
-
-	var err error
-	if !allowedGitProviders[gitProvider] {
-		// Self-hosted git provider, check for git-provider annotation on the component
-		gitProviderAnnotationValue := component.GetAnnotations()[GitProviderAnnotationName]
-		if gitProviderAnnotationValue != "" {
-			if allowedGitProviders[gitProviderAnnotationValue] {
-				gitProvider = gitProviderAnnotationValue
-			} else {
-				err = fmt.Errorf("unsupported \"%s\" annotation value: %s", GitProviderAnnotationName, gitProviderAnnotationValue)
-			}
-		} else {
-			err = fmt.Errorf("self-hosted git provider is not specified via \"%s\" annotation in the component", GitProviderAnnotationName)
-		}
-	}
-
-	return gitProvider, err
 }
 
 // SetDefaultBuildPipelineComponentAnnotation sets default build pipeline to component pipeline annotation
@@ -185,6 +133,24 @@ func (r *ComponentBuildReconciler) GetBuildPipelineFromComponentAnnotation(ctx c
 	return pipelineRef, nil
 }
 
+func readBuildPipelineAnnotation(component *appstudiov1alpha1.Component) (*BuildPipeline, error) {
+	if component.Annotations == nil {
+		return nil, nil
+	}
+
+	requestedPipeline, requestedPipelineExists := component.Annotations[defaultBuildPipelineAnnotation]
+	if requestedPipelineExists && requestedPipeline != "" {
+		buildPipeline := &BuildPipeline{}
+		buildPipelineBytes := []byte(requestedPipeline)
+
+		if err := json.Unmarshal(buildPipelineBytes, buildPipeline); err != nil {
+			return nil, boerrors.NewBuildOpError(boerrors.EFailedToParsePipelineAnnotation, err)
+		}
+		return buildPipeline, nil
+	}
+	return nil, nil
+}
+
 func (r *ComponentBuildReconciler) ensurePipelineServiceAccount(ctx context.Context, namespace string) (*corev1.ServiceAccount, error) {
 	log := ctrllog.FromContext(ctx)
 
@@ -249,97 +215,4 @@ func getComponentImageRepoAndSecretNameFromImageAnnotation(component *appstudiov
 		return repoInfo.Image, repoInfo.Secret, nil
 	}
 	return "", "", nil
-}
-
-// mergeAndSortTektonParams merges additional params into existing params by adding new or replacing existing values.
-func mergeAndSortTektonParams(existedParams, additionalParams []tektonapi.Param) []tektonapi.Param {
-	var params []tektonapi.Param
-	paramsMap := make(map[string]tektonapi.Param)
-	for _, p := range existedParams {
-		paramsMap[p.Name] = p
-	}
-	for _, p := range additionalParams {
-		paramsMap[p.Name] = p
-	}
-	for _, v := range paramsMap {
-		params = append(params, v)
-	}
-	sort.Slice(params, func(i, j int) bool {
-		return params[i].Name < params[j].Name
-	})
-	return params
-}
-
-func generateVolumeClaimTemplate() *corev1.PersistentVolumeClaim {
-	return &corev1.PersistentVolumeClaim{
-		Spec: corev1.PersistentVolumeClaimSpec{
-			AccessModes: []corev1.PersistentVolumeAccessMode{
-				"ReadWriteOnce",
-			},
-			Resources: corev1.VolumeResourceRequirements{
-				Requests: corev1.ResourceList{
-					"storage": resource.MustParse("1Gi"),
-				},
-			},
-		},
-	}
-}
-
-func getPathContext(gitContext, dockerfileContext string) string {
-	if gitContext == "" && dockerfileContext == "" {
-		return ""
-	}
-	separator := string(filepath.Separator)
-	path := filepath.Join(gitContext, dockerfileContext)
-	path = filepath.Clean(path)
-	path = strings.TrimPrefix(path, separator)
-	return path
-}
-
-func getPipelineNameAndBundle(pipelineRef *tektonapi.PipelineRef) (string, string, error) {
-	if pipelineRef.Resolver != "" && pipelineRef.Resolver != "bundles" {
-		return "", "", boerrors.NewBuildOpError(
-			boerrors.EUnsupportedPipelineRef,
-			fmt.Errorf("unsupported Tekton resolver %q", pipelineRef.Resolver),
-		)
-	}
-
-	name := pipelineRef.Name
-	var bundle string
-
-	for _, param := range pipelineRef.Params {
-		switch param.Name {
-		case "name":
-			name = param.Value.StringVal
-		case "bundle":
-			bundle = param.Value.StringVal
-		}
-	}
-
-	if name == "" || bundle == "" {
-		return "", "", boerrors.NewBuildOpError(
-			boerrors.EMissingParamsForBundleResolver,
-			fmt.Errorf("missing name or bundle in pipelineRef: name=%s bundle=%s", name, bundle),
-		)
-	}
-
-	return name, bundle, nil
-}
-
-func readBuildPipelineAnnotation(component *appstudiov1alpha1.Component) (*BuildPipeline, error) {
-	if component.Annotations == nil {
-		return nil, nil
-	}
-
-	requestedPipeline, requestedPipelineExists := component.Annotations[defaultBuildPipelineAnnotation]
-	if requestedPipelineExists && requestedPipeline != "" {
-		buildPipeline := &BuildPipeline{}
-		buildPipelineBytes := []byte(requestedPipeline)
-
-		if err := json.Unmarshal(buildPipelineBytes, buildPipeline); err != nil {
-			return nil, boerrors.NewBuildOpError(boerrors.EFailedToParsePipelineAnnotation, err)
-		}
-		return buildPipeline, nil
-	}
-	return nil, nil
 }
