+++ controllers/component_build_controller_unit_test.go
@@ -17,7 +17,6 @@ limitations under the License.
 package controllers
 
 import (
-	"context"
 	"encoding/base64"
 	"fmt"
 	"reflect"
@@ -196,199 +195,6 @@ func TestWriteBuildStatus(t *testing.T) {
 				t.Errorf("writeBuildStatus(): actual: %v, want %v", got, tt.want)
 			}
 		})
-	}
-}
-
-func TestGenerateInitialPipelineRunForComponentDockerfileContext(t *testing.T) {
-	component := &appstudiov1alpha1.Component{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "my-component",
-			Namespace: "my-namespace",
-			Annotations: map[string]string{
-				"skip-initial-checks":          "true",
-				GitProviderAnnotationName:      "github",
-				defaultBuildPipelineAnnotation: testPipelineAnnotation,
-			},
-		},
-		Spec: appstudiov1alpha1.ComponentSpec{
-			Application:    "my-application",
-			ContainerImage: "registry.io/username/image:tag",
-			Source: appstudiov1alpha1.ComponentSource{
-				ComponentSourceUnion: appstudiov1alpha1.ComponentSourceUnion{
-					GitSource: &appstudiov1alpha1.GitSource{
-						URL:           "https://githost.com/user/repo.git",
-						Revision:      "custom-branch",
-						Context:       "./base_context",
-						DockerfileURL: "containerFile",
-					},
-				},
-			},
-		},
-		Status: appstudiov1alpha1.ComponentStatus{},
-	}
-	pipelineRef := &tektonapi.PipelineRef{
-		ResolverRef: tektonapi.ResolverRef{
-			Resolver: "bundles",
-			Params: []tektonapi.Param{
-				{Name: "name", Value: *tektonapi.NewStructuredValues("pipeline-name")},
-				{Name: "bundle", Value: *tektonapi.NewStructuredValues("pipeline-bundle")},
-			},
-		},
-	}
-
-	pipelineRun, err := generatePipelineRunForComponent(component, pipelineRef, &buildGitInfo{})
-
-	if err != nil {
-		t.Error("generateInitialPipelineRunForComponentDockerfileContext(): Failed to generate pipeline run")
-	}
-
-	for _, param := range pipelineRun.Spec.Params {
-		switch param.Name {
-		case "path-context":
-			if param.Value.StringVal != "base_context" {
-				t.Errorf("generateInitialPipelineRunForComponentDockerfileContext(): wrong pipeline parameter %s", param.Name)
-			}
-		case "dockerfile":
-			if param.Value.StringVal != "containerFile" {
-				t.Errorf("generateInitialPipelineRunForComponentDockerfileContext(): wrong pipeline parameter %s", param.Name)
-			}
-		case "revision":
-			if param.Value.StringVal != "custom-branch" {
-				t.Errorf("generateInitialPipelineRunForComponentDockerfileContext(): wrong pipeline parameter %s", param.Name)
-			}
-		}
-	}
-}
-
-func TestGenerateInitialPipelineRunForComponent(t *testing.T) {
-	component := &appstudiov1alpha1.Component{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "my-component",
-			Namespace: "my-namespace",
-			Annotations: map[string]string{
-				"skip-initial-checks":          "true",
-				GitProviderAnnotationName:      "github",
-				defaultBuildPipelineAnnotation: testPipelineAnnotation,
-			},
-		},
-		Spec: appstudiov1alpha1.ComponentSpec{
-			Application:    "my-application",
-			ContainerImage: "registry.io/username/image:tag",
-			Source: appstudiov1alpha1.ComponentSource{
-				ComponentSourceUnion: appstudiov1alpha1.ComponentSourceUnion{
-					GitSource: &appstudiov1alpha1.GitSource{
-						URL:      "https://githost.com/user/repo.git",
-						Revision: "custom-branch",
-						Context:  "./base_context",
-					},
-				},
-			},
-		},
-		Status: appstudiov1alpha1.ComponentStatus{},
-	}
-	pipelineRef := &tektonapi.PipelineRef{
-		ResolverRef: tektonapi.ResolverRef{
-			Resolver: "bundles",
-			Params: []tektonapi.Param{
-				{Name: "name", Value: *tektonapi.NewStructuredValues("pipeline-name")},
-				{Name: "bundle", Value: *tektonapi.NewStructuredValues("pipeline-bundle")},
-			},
-		},
-	}
-	commitSha := "26239c94569cea79b32bce32f12c8abd8bbd0fd7"
-
-	pRunGitInfo := &buildGitInfo{
-		gitSourceSha:              commitSha,
-		browseRepositoryAtShaLink: "https://githost.com/user/repo?rev=" + commitSha,
-	}
-
-	pipelineRun, err := generatePipelineRunForComponent(component, pipelineRef, pRunGitInfo)
-	if err != nil {
-		t.Error("generateInitialPipelineRunForComponent(): Failed to genertate pipeline run")
-	}
-
-	if pipelineRun.GenerateName == "" {
-		t.Error("generateInitialPipelineRunForComponent(): pipeline generatename must not be empty")
-	}
-	if pipelineRun.Namespace != "my-namespace" {
-		t.Error("generateInitialPipelineRunForComponent(): pipeline namespace doesn't match")
-	}
-
-	if pipelineRun.Labels[ApplicationNameLabelName] != "my-application" {
-		t.Errorf("generateInitialPipelineRunForComponent(): wrong %s label value", ApplicationNameLabelName)
-	}
-	if pipelineRun.Labels[ComponentNameLabelName] != "my-component" {
-		t.Errorf("generateInitialPipelineRunForComponent(): wrong %s label value", ComponentNameLabelName)
-	}
-	if pipelineRun.Labels["pipelines.appstudio.openshift.io/type"] != "build" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong pipelines.appstudio.openshift.io/type label value")
-	}
-
-	if pipelineRun.Annotations["build.appstudio.redhat.com/target_branch"] != "custom-branch" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong build.appstudio.redhat.com/target_branch annotation value")
-	}
-	if pipelineRun.Annotations["build.appstudio.redhat.com/pipeline_name"] != "pipeline-name" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong build.appstudio.redhat.com/pipeline_name annotation value")
-	}
-	if pipelineRun.Annotations["build.appstudio.redhat.com/bundle"] != "pipeline-bundle" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong build.appstudio.redhat.com/bundle annotation value")
-	}
-	if pipelineRun.Annotations[gitCommitShaAnnotationName] != commitSha {
-		t.Errorf("generateInitialPipelineRunForComponent(): wrong %s annotation value", gitCommitShaAnnotationName)
-	}
-	if pipelineRun.Annotations[gitRepoAtShaAnnotationName] != "https://githost.com/user/repo?rev="+commitSha {
-		t.Errorf("generateInitialPipelineRunForComponent(): wrong %s annotation value", gitRepoAtShaAnnotationName)
-	}
-
-	if getPipelineName(pipelineRun.Spec.PipelineRef) != "pipeline-name" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong pipeline name in pipeline reference")
-	}
-	if getPipelineBundle(pipelineRun.Spec.PipelineRef) != "pipeline-bundle" {
-		t.Error("generateInitialPipelineRunForComponent(): wrong pipeline bundle in pipeline reference")
-	}
-
-	if len(pipelineRun.Spec.Params) != 6 {
-		t.Error("generateInitialPipelineRunForComponent(): wrong number of pipeline params")
-	}
-	for _, param := range pipelineRun.Spec.Params {
-		switch param.Name {
-		case "git-url":
-			if param.Value.StringVal != "https://githost.com/user/repo.git" {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s", param.Name)
-			}
-		case "revision":
-			if param.Value.StringVal != commitSha {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s value", param.Name)
-			}
-		case "output-image":
-			if !strings.HasPrefix(param.Value.StringVal, "registry.io/username/image:build-") {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s value", param.Name)
-			}
-		case "skip-checks":
-			if param.Value.StringVal != "true" {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s value", param.Name)
-			}
-		case "dockerfile":
-			if param.Value.StringVal != "Dockerfile" {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s value", param.Name)
-			}
-		case "path-context":
-			if param.Value.StringVal != "base_context" {
-				t.Errorf("generateInitialPipelineRunForComponent(): wrong pipeline parameter %s value", param.Name)
-			}
-		default:
-			t.Errorf("generateInitialPipelineRunForComponent(): unexpected pipeline parameter %v", param)
-		}
-	}
-
-	if len(pipelineRun.Spec.Workspaces) != 1 {
-		t.Error("generateInitialPipelineRunForComponent(): wrong number of pipeline workspaces")
-	}
-	for _, workspace := range pipelineRun.Spec.Workspaces {
-		if workspace.Name == "workspace" {
-			continue
-		}
-		t.Errorf("generateInitialPipelineRunForComponent(): unexpected pipeline workspaces %v", workspace)
 	}
 }
 
@@ -687,193 +493,6 @@ func TestGetContainerImageRepository(t *testing.T) {
 			got := getContainerImageRepository(tt.image)
 			if got != tt.want {
 				t.Errorf("getContainerImageRepository(): got %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestMergeAndSortTektonParams(t *testing.T) {
-	tests := []struct {
-		name       string
-		existing   []tektonapi.Param
-		additional []tektonapi.Param
-		want       []tektonapi.Param
-	}{
-		{
-			name: "should merge two different parameters lists",
-			existing: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-			additional: []tektonapi.Param{
-				{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: "docker/Dockerfile"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-			},
-			want: []tektonapi.Param{
-				{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: "docker/Dockerfile"}},
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-		},
-		{
-			name: "should append empty parameters list",
-			existing: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-			additional: []tektonapi.Param{},
-			want: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-		},
-		{
-			name: "should sort parameters list",
-			existing: []tektonapi.Param{
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-				{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: "docker/Dockerfile"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-			},
-			additional: []tektonapi.Param{},
-			want: []tektonapi.Param{
-				{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: "docker/Dockerfile"}},
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-		},
-		{
-			name: "should override existing parameters",
-			existing: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "false"}},
-			},
-			additional: []tektonapi.Param{
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "2378a064bf6b66a8ffc650ad88d404cca24ade29"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-			},
-			want: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "2378a064bf6b66a8ffc650ad88d404cca24ade29"}},
-			},
-		},
-		{
-			name: "should append and override parameters",
-			existing: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "main"}},
-			},
-			additional: []tektonapi.Param{
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "2378a064bf6b66a8ffc650ad88d404cca24ade29"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-			},
-			want: []tektonapi.Param{
-				{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "https://githost.com/user/repo.git"}},
-				{Name: "rebuild", Value: tektonapi.ParamValue{Type: "string", StringVal: "true"}},
-				{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "2378a064bf6b66a8ffc650ad88d404cca24ade29"}},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := mergeAndSortTektonParams(tt.existing, tt.additional)
-			if len(got) != len(tt.want) {
-				t.Errorf("mergeAndSortTektonParams(): got %v, want %v", got, tt.want)
-			}
-			for i := range got {
-				if !reflect.DeepEqual(got[i], tt.want[i]) {
-					t.Errorf("mergeAndSortTektonParams(): got %v, want %v", got, tt.want)
-				}
-			}
-		})
-	}
-}
-
-func TestGetGitProviderUrl(t *testing.T) {
-	type args struct {
-		ctx    context.Context
-		gitURL string
-	}
-	tests := []struct {
-		name       string
-		args       args
-		wantErr    bool
-		wantString string
-	}{
-		{
-			name: "github ssh",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "git@github.com:redhat-appstudio/application-service.git",
-			},
-			wantErr:    false,
-			wantString: "https://github.com",
-		},
-		{
-			name: "gitlab ssh",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "git@gitlab.com:namespace/project-name.git",
-			},
-			wantErr:    false,
-			wantString: "https://gitlab.com",
-		},
-		{
-			name: "bitbucket ssh",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "git@bitbucket.org:organization/project-name.git",
-			},
-			wantErr:    false,
-			wantString: "https://bitbucket.org",
-		},
-		{
-			name: "github https",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "https://github.com/redhat-appstudio/application-service.git",
-			},
-			wantErr:    false,
-			wantString: "https://github.com",
-		},
-		{
-			name: "bitbucket https",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "https://sbose78@bitbucket.org/sbose78/appstudio.git",
-			},
-			wantErr:    false,
-			wantString: "https://bitbucket.org",
-		},
-		{
-			name: "no scheme",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "github.com/redhat-appstudio/application-service.git",
-			},
-			wantErr:    true, //fully qualified URL is a must
-			wantString: "",
-		},
-		{
-			name: "invalid url",
-			args: args{
-				ctx:    context.Background(),
-				gitURL: "not-even-a-url",
-			},
-			wantErr:    true, //fully qualified URL is a must
-			wantString: "",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got, err := getGitProviderUrl(tt.args.gitURL); (got != tt.wantString) ||
-				(tt.wantErr == true && err == nil) ||
-				(tt.wantErr == false && err != nil) {
-				t.Errorf("getGitProviderUrl() Got Error: = %v, want %v ; Got String:  = %v , want %v", err, tt.wantErr, got, tt.wantString)
 			}
 		})
 	}
